restructure code for tests
