+++ python/fastquant/fastquant.py
@@ -8,7 +8,8 @@ Created on Tue Jun 25 19:48:03 2019
 # Import standard library
 import os
 import requests
-from datetime import datetime
+from datetime import datetime, timedelta
+import time
 from pathlib import Path
 from pkg_resources import resource_filename
 
@@ -19,6 +20,10 @@ import lxml.html as LH
 from tqdm import tqdm
 import tweepy
 import yfinance as yf
+from nltk.sentiment.vader import SentimentIntensityAnalyzer
+import nltk
+from urllib.request import urlopen
+from bs4 import BeautifulSoup
 
 DATA_PATH = resource_filename(__name__, "data")
 
@@ -567,3 +572,69 @@ def datestring_to_datetime(date, sep="-"):
     errmsg = "date format must be YYYY-MM-DD"
     assert len(ymd[0]) == 4, errmsg
     return datetime(*map(int, ymd))
+
+
+def get_bt_news(keyword=None, page_nums=None):
+    """
+    This function scrapes Business Times (https://www.businesstimes.com.sg) articles by giving
+    a specific keyword e.g "facebook, jollibee" and number of pages that you needed.
+
+    Parameters
+    ----------
+    keyword : str
+        The keyword you wanted to search for in Business Times page.
+    page_nums : int
+        The number of iteration of pages you want to scrape.
+
+    Returns
+    ----------
+    date_sentiments: dict
+        The dictionary output of the scraped data in form of {date: sentiment score}
+    """
+
+    nltk.download("vader_lexicon", quiet=True)  # download vader lexicon
+
+    # Check for keyword
+    if keyword is None:
+        print(
+            "keyword is set to None, please type in keyword or company name that you wanted to scrape e.g jollibee, facebook etc."
+        )
+        raise TypeError
+    if page_nums is None:
+        page_nums = 1
+        print("no page numbers indicated, setting this variable to 1")
+
+    date_sentiments = {}
+    sia = SentimentIntensityAnalyzer()
+
+    for i in tqdm(range(1, page_nums + 1)):
+        page = urlopen(
+            f"https://www.businesstimes.com.sg/search/{keyword}?page={str(i)}"
+        ).read()
+        soup = BeautifulSoup(page, features="html.parser")
+        posts = soup.findAll("div", {"class": "media-body"})
+        for post in posts:
+            time.sleep(1)
+            url = post.a["href"]
+            date = post.time.text
+            try:
+                link_page = urlopen(url).read()
+            except Exception:
+                url = url[:-2]
+                link_page = urlopen(url).read()
+            link_soup = BeautifulSoup(link_page, features="lxml")
+            sentences = link_soup.findAll("p")
+            passage = ""
+            for sentence in sentences:
+                passage += sentence.text
+            sentiment = sia.polarity_scores(passage)["compound"]
+            date_sentiments.setdefault(date, []).append(sentiment)
+
+    date_sentiment = {}
+
+    for k, v in date_sentiments.items():
+        date_sentiment[
+            datetime.strptime(k, "%d %b %Y").date() + timedelta(days=1)
+        ] = round(sum(v) / float(len(v)), 3)
+
+    return date_sentiment
